app/requirements.txt
click==8.0.4
Flask==2.0.3
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
paho-mqtt==1.6.1
python-dotenv==0.19.2
Werkzeug==2.0.3
